refactor(ios): move avatar cache infrastructure to Core (#14254)
## Summary

Iteration 4 of the iOS refactor loop (no Linear issue — ad-hoc loop work). Behavior-preserving split of `Features/Dashboard/DashboardView.swift` (**496 → 388 lines**).

- `AvatarImageCache`, `AvatarImageLoader`, and `CachedRemoteImageView` move **verbatim** to a new `Core/AvatarImageCache.swift`. These are the sanctioned cache-first image loading primitives referenced by `.claude/rules/ios.md` (`no-raw-asyncimage` remediation) and are shared by dashboard avatars and chat entity chips — Core is their right home.
- `DashboardAvatarView` (dashboard-specific) stays in the feature file.
- Names kept stable; `project.pbxproj` wired with explicit entries (`plutil -lint` clean).

## Verification

- `pnpm run ios:lint` → **clean ✓**
- `pnpm run ios:test` (iPhone 17 sim, iOS 26.3.1) → **248 passed, 6 skipped, 2 failed** — the failures are the known local-sim-only pair tracked in **JOV-4252** (pass on GitHub CI; identical on clean `main`).

## Risks

- Low: verbatim same-module move; no access-level or name changes.
